requirements.txt
Django>=4.1.3
pytz>=2022.6
pyYAML>=6.0
ipython>=7.34.0
djangorestframework==3.14.0
google-api-core==2.14.0
google-auth==2.24.0
google-cloud-compute==1.14.1
googleapis-common-protos==1.61.0
proto-plus==1.22.3
protobuf==4.25.1
django-cors-headers==4.3.1
pycryptodome==3.19.0